Fix all clippy warnings across workspace (25 files)
- Convert /// doc comments to //! module-level doc comments in all crate
  modules to fix empty_line_after_doc_comments warnings (~20 files)
- Replace .map_or() with .is_some_and() in lexer.rs and install.rs
- Add Default impl for TypeChecker (checker.rs) and Installer (install.rs)
- Replace match single-pattern with if-let in optimizer.rs
- Remove unnecessary cast (then_len as i32) in bytecode_gen.rs
- Replace &[item.clone()] with std::slice::from_ref in bootstrap.rs
- Fix doc_lazy_continuation by adding blank line before struct doc

Result: 0 clippy warnings across all 6 crates

+++ crates/whisper-codegen/src/bytecode_gen.rs
@@ -1,8 +1,8 @@
-/// Bytecode generator: AST → Vec<Opcode>.
-///
-/// Two-pass compilation:
-///   Pass 1: Pre-scan all Def nodes, compile their bodies, build word_dict.
-///   Pass 2: Compile main program, resolving WordRef by inlining from word_dict.
+//! Bytecode generator: AST → Vec<Opcode>.
+//!
+//! Two-pass compilation:
+//!   Pass 1: Pre-scan all Def nodes, compile their bodies, build word_dict.
+//!   Pass 2: Compile main program, resolving WordRef by inlining from word_dict.
 
 use std::collections::HashMap;
 use whisper_core::opcode::Opcode;
@@ -186,7 +186,7 @@ impl BytecodeGenerator {
             self.emit(Opcode::Jump(jmp_off));
             self.bytecode.extend(else_c.clone());
         } else {
-            self.emit(Opcode::Cond(then_len as i32));
+            self.emit(Opcode::Cond(then_len));
             self.bytecode.extend(then_code);
         }
     }
